Fix PR review findings
